# public/openapi.yaml
...
openapi: 3.1.0
info:
  title: Card Orb API
  version: 1.0.0
  summary: A Pokémon card collection, priced, served to a web tool and an iOS app.
  description: |
    Card Orb keeps one person's Pokémon card collection in Postgres, matches it
    against three card catalogues, prices it from Cardmarket, and serves it here.

    **Two clients read this API and both are ours**: the web tool at cardorb.com
    and the iOS app. Nothing here is designed for a third party yet.

    **Every failure is `{ "error": "<sentence>" }`** at a status a client can
    branch on. The sentence is for a person to read; the status is for the code.

    **The version is in the path.** `/v1` changes only by addition: a new field,
    a new route, a new optional query. Anything that removes or renames is `/v2`.
  contact:
    name: Card Orb
    url: https://cardorb.com
servers:
  - url: https://api.cardorb.com
    description: The API on its own host. `/v1/…` here is `/api/v1/…` on cardorb.com.
  - url: https://cardorb.com/api
    description: The same API under the web tool's own origin, which is what the browser uses.
  - url: http://localhost:3000/api
    description: Local development.
...
security:
  - bearer: []
  - session: []
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A Supabase access token, `Authorization: Bearer <jwt>`. What the iOS app
        sends. Verified locally against the project's signing keys.
    session:
      type: apiKey
      in: cookie
      name: binder_session
      description: |
        The browser's session cookie, set by `POST /v1/session`. What the web
        tool sends. Scoped to cardorb.com, so it never reaches api.cardorb.com —
        the browser calls `/api/v1` on its own origin instead.
    cron:
      type: http
      scheme: bearer
      description: Vercel's `CRON_SECRET`, on `/v1/cron/snapshot` only.
...
  responses:
    BadRequest:
      description: The request was understood and refused. The sentence says why.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Invalid request" }
    Unauthorised:
      description: No viewer. Sign in, or send a bearer token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Sign in to see this." }
    Forbidden:
      description: The request came from an origin this deployment does not allow.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Forbidden" }
...
    PayloadTooLarge:
      description: The body is bigger than this route's `BODY_LIMIT`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Payload too large" }
...
    TooManyRequests:
      description: Over this route's rate limit for this address.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Too many requests" }
    BadGateway:
      description: A catalogue, the price guide or the database did not answer.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unavailable:
      description: This deployment has no database, or the database is not answering.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "This deployment has no database configured." }
  schemas:
    Error:
      type: object
      description: The one shape every failure takes.
      required: [error]
      properties:
        error:
          type: string
          description: A sentence for a person. Branch on the status, not on this.
      additionalProperties: true
...
paths:
...
  /v1/username:
    post:
      tags: [Profile]
      operationId: claimUsername
      summary: Claim a username
      description: Same-origin only. Lower-cased and trimmed first; claiming your own name is a no-op that answers 200.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username: { type: string }
      responses:
        "200":
          description: Claimed.
          content:
            application/json:
              schema:
                type: object
                required: [ok, username]
                properties:
                  ok: { type: boolean, const: true }
                  username: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorised" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Taken, or reserved.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "500": { $ref: "#/components/responses/BadGateway" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
  /v1/signup:
    post:
      tags: [Account]
      operationId: signUp
      summary: Create an account
      description: Same-origin only. Five per address each fifteen minutes. Sends a confirmation email; the account is `pending` until it is followed.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 10 }
                name: { type: string, maxLength: 60 }
      responses:
        "200":
          description: Created, pending confirmation.
          content:
            application/json:
              schema:
                type: object
                required: [ok, pending, email]
                properties:
                  ok: { type: boolean, const: true }
                  pending: { type: boolean, const: true }
                  email: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403":
          description: Wrong origin, or new accounts are closed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: That address is already in use.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/Unavailable" }
...